+++ HISTORY.md
@@ -11,3 +11,4 @@ you spot any mistakes.
 * We're doing better escaping of identifiers now
 * Added the ability to use table/view prefixes directly (ex. `select * from users u`)
 * Added the ability to do `OR WHERE` statements with `or_where()` method.
+* Added the ability to do `LIKE` statements directly (new methods: `like()`, `not_like()`, `or_like()`, `or_not_like()`)